Initial Commit
Getting a baseline down with placeholders so that when I have access to the database I can just plug in website addresses and class names and be running quickly.
TO DO: get Pulses vs DM running, Noise vs DM, and Correlation. Furthermore get already determined data to train the AI model to be able to detect pulsars.